+++ src/react/index.ts
@@ -1,4 +1,8 @@
 export { default as VisualizeWidgetFrame } from './VisualizeWidgetFrame'
-export type { VisualizeWidgetFrameProps } from './VisualizeWidgetFrame'
+export type {
+  StreamVisualizationTheme,
+  StreamVisualizationThemeTokens,
+  VisualizeWidgetFrameProps,
+} from './VisualizeWidgetFrame'
 export { default as StreamVisualization } from './VisualizeWidgetFrame'
 export type { VisualizeWidgetFrameProps as StreamVisualizationProps } from './VisualizeWidgetFrame'